Time: 76 ms (69.75%) | Memory: 117.3 MB (7.56%) - LeetSync

+++ 883-car-fleet/car-fleet.java
@@ -11,7 +11,7 @@ class Solution {
       for(int i= position.length-1;i>=0;i--){
         double dis=target-position[i];
         double sp=map.get(position[i]);
-            double t=(double)(dis/sp);
+            double t=(dis/sp);
             if(t>max){
                 count++;
                 max=t;
